updated pagination and grouping logic
